@@ -1,6 +1,6 @@
 {
-	"name": "nette/sandbox",
-	"description": "The sandbox is a pre-packaged Nette Framework project, basic configured structure for your application.",
+	"name": "Texila",
+	"description": "The sandbox for Texila.",
 	"homepage": "http://nette.org",
 	"type": "project",
 	"license": ["BSD-3-Clause", "GPL-2.0", "GPL-3.0"],
@@ -15,13 +15,27 @@
 		}
 	],
 	"require": {
-		"php": ">= 5.3.7",
-		"nette/nette": "v2.1.1",
-		"dg/adminer-custom": "1.*",
-        "dg/texy": "v2.6",
-        "JanMarek/WebLoader": "dev-jm-nette-extension"
+		"php": ">= 5.4.0",
+		"nette/application": "~2.3.6",
+		"nette/bootstrap": "~2.3.0",
+		"nette/caching": "~2.3.0",
+		"nette/database": "~2.3.0",
+		"nette/di": "~2.3.0",
+		"nette/finder": "~2.3.0",
+		"nette/forms": "~2.3.0",
+		"nette/http": "~2.3.0",
+		"nette/mail": "~2.3.0",
+		"nette/robot-loader": "~2.3.0",
+		"nette/safe-stream": "~2.3.0",
+		"nette/security": "~2.3.0",
+		"nette/utils": "~2.3.0",
+		"latte/latte": "~2.3.0",
+		"tracy/tracy": "~2.3.0",
+		"dg/adminer-custom": "~1.6",
+    "dg/texy": "2.8.*",
+    "JanMarek/WebLoader": "^2.5"
 	},
 	"require-dev": {
-		"nette/tester": "~1.0.0"
+		"nette/tester": "~1.3"
 	}
 }
