@@ -2,50 +2,37 @@
 
 namespace App\Presenters;
 
-use Nette,
-	App\Model;
-
+use Nette;
 
 /**
  * Homepage presenter.
  */
-class HomepagePresenter extends BasePresenter
-{
+class HomepagePresenter extends BasePresenter {
+  
+  /**
+   * Example form factory
+   * @return Form
+   */
+  protected function createComponentExampleForm() {
+    $form = new Nette\Application\UI\Form;
 
-    /**
-     * Example form factory
-     * @return Form
-     */
-    protected function createComponentExampleForm()
-    {
-        $form = new Nette\Application\UI\Form;
+    $form->addTextarea("text", NULL)
+        ->setAttribute('cols', 80)
+        ->setAttribute('rows', 20)
+        ->getControlPrototype()->class("texyla");
 
-        $form->addTextarea("text", NULL)
-            ->setAttribute('cols', 80)
-            ->setAttribute('rows', 20)
-            ->getControlPrototype()->class("texyla");
-
-        $form->addSubmit("s", "Submit");
-        $form->onSuccess[] = callback($this, 'exampleFormSubmitted');
-        return $form;
-    }
-
+    $form->addSubmit("s", "Submit");
+    $form->onSuccess[] = $this->exampleFormSubmitted;
+    return $form;
+  }
 	
-	public function exampleFormSubmitted($form){
+	public function exampleFormSubmitted($form) {
+    $values = $form->getValues();
+    dump($values);
+  }
 
-        $values = $form->getValues();
-
-        dump($values);
-
-        //add redirect later 
-
-    }
-
-	public function renderDefault()
-	{
-
-
-        $this->template->texytest = '
+	public function renderDefault() {
+    $this->template->texytest = '
         Vítejte!
 --------
 
@@ -58,16 +45,15 @@ Můžete používat syntax Texy!, pokud Vám vyhovuje:
 Ale také můžete zůstat u HTML:
 - takto <b>HTML</b>
 - nebo i <b class=xx>úplně <i>hloupě</b>, Texy! to pořeší';
-
 	}
 
 
-    protected function createTemplate($class = NULL) {
-        $template = parent::createTemplate($class);
-        $texy = new \Texy();
-        $texy->allowedTags = false;
-        $template->registerHelper('texy', callback($texy, 'process'));
-        return $template;
-    }
+  protected function createTemplate($class = NULL) {
+    $template = parent::createTemplate($class);
+    $this->texy->allowedTags = false;
+    $this->texy->headingModule->balancing = "FIXED";
+    $template->addFilter('texy', [$this->texy, 'process']);
+    return $template;
+  }
 
 }
